Move computed value logic to trait
